Merge release 10.2.4

+++ submodules/TelegramCore/Sources/AccountManager/AccountManagerImpl.swift
@@ -82,6 +82,7 @@ final class AccountManagerImpl<Types: AccountManagerTypes> {
             return (records, currentRecordId)
         } catch let e {
             postboxLog("decode atomic state error: \(e)")
+            postboxLogSync()
             preconditionFailure()
         }
     }
@@ -96,11 +97,29 @@ final class AccountManagerImpl<Types: AccountManagerTypes> {
         self.temporarySessionId = temporarySessionId
         let _ = try? FileManager.default.createDirectory(atPath: basePath, withIntermediateDirectories: true, attributes: nil)
         guard let guardValueBox = SqliteValueBox(basePath: basePath + "/guard_db", queue: queue, isTemporary: isTemporary, isReadOnly: false, useCaches: useCaches, removeDatabaseOnError: removeDatabaseOnError, encryptionParameters: nil, upgradeProgress: { _ in }) else {
+            postboxLog("Could not open guard value box at \(basePath + "/guard_db")")
+            postboxLogSync()
+            preconditionFailure()
             return nil
         }
         self.guardValueBox = guardValueBox
-        guard let valueBox = SqliteValueBox(basePath: basePath + "/db", queue: queue, isTemporary: isTemporary, isReadOnly: isReadOnly, useCaches: useCaches, removeDatabaseOnError: removeDatabaseOnError, encryptionParameters: nil, upgradeProgress: { _ in }) else {
-            return nil
+        
+        var valueBox: SqliteValueBox?
+        for i in 0 ..< 3 {
+            if let valueBoxValue = SqliteValueBox(basePath: basePath + "/db", queue: queue, isTemporary: isTemporary, isReadOnly: isReadOnly, useCaches: useCaches, removeDatabaseOnError: removeDatabaseOnError, encryptionParameters: nil, upgradeProgress: { _ in }) {
+                valueBox = valueBoxValue
+                break
+            } else {
+                postboxLog("Could not open value box at \(basePath + "/db") (try \(i))")
+                postboxLogSync()
+                
+                Thread.sleep(forTimeInterval: 0.1 + 0.5 * Double(i))
+            }
+        }
+        guard let valueBox = valueBox else {
+            postboxLog("Giving up on opening value box at \(basePath + "/db")")
+            postboxLogSync()
+            preconditionFailure()
         }
         self.valueBox = valueBox
         
@@ -116,17 +135,27 @@ final class AccountManagerImpl<Types: AccountManagerTypes> {
                 self.currentAtomicState = atomicState
             } catch let e {
                 postboxLog("decode atomic state error: \(e)")
-                let _ = try? FileManager.default.removeItem(atPath: self.atomicStatePath)
+                postboxLogSync()
+                
+                if removeDatabaseOnError {
+                    let _ = try? FileManager.default.removeItem(atPath: self.atomicStatePath)
+                }
                 preconditionFailure()
             }
         } catch let e {
             postboxLog("load atomic state error: \(e)")
-            var legacyRecordDict: [AccountRecordId: AccountRecord<Types.Attribute>] = [:]
-            for record in self.legacyRecordTable.getRecords() {
-                legacyRecordDict[record.id] = record
+            postboxLogSync()
+            
+            if removeDatabaseOnError {
+                var legacyRecordDict: [AccountRecordId: AccountRecord<Types.Attribute>] = [:]
+                for record in self.legacyRecordTable.getRecords() {
+                    legacyRecordDict[record.id] = record
+                }
+                self.currentAtomicState = AccountManagerAtomicState(records: legacyRecordDict, currentRecordId: self.legacyMetadataTable.getCurrentAccountId(), currentAuthRecord: self.legacyMetadataTable.getCurrentAuthAccount(), accessChallengeData: self.legacyMetadataTable.getAccessChallengeData())
+                self.syncAtomicStateToFile()
+            } else {
+                preconditionFailure()
             }
-            self.currentAtomicState = AccountManagerAtomicState(records: legacyRecordDict, currentRecordId: self.legacyMetadataTable.getCurrentAccountId(), currentAuthRecord: self.legacyMetadataTable.getCurrentAuthAccount(), accessChallengeData: self.legacyMetadataTable.getAccessChallengeData())
-            self.syncAtomicStateToFile()
         }
         
         let tableAccessChallengeData = self.legacyMetadataTable.getAccessChallengeData()
@@ -267,9 +296,11 @@ final class AccountManagerImpl<Types: AccountManagerTypes> {
         if let data = try? JSONEncoder().encode(self.currentAtomicState) {
             if let _ = try? data.write(to: URL(fileURLWithPath: self.atomicStatePath), options: [.atomic]) {
             } else {
+                postboxLogSync()
                 preconditionFailure()
             }
         } else {
+            postboxLogSync()
             preconditionFailure()
         }
     }
@@ -617,10 +648,11 @@ public final class AccountManager<Types: AccountManagerTypes> {
             if let value = AccountManagerImpl<Types>(queue: queue, basePath: basePath, isTemporary: isTemporary, isReadOnly: isReadOnly, useCaches: useCaches, removeDatabaseOnError: removeDatabaseOnError, temporarySessionId: temporarySessionId) {
                 return value
             } else {
+                postboxLogSync()
                 preconditionFailure()
             }
         })
-        self.mediaBox = MediaBox(basePath: basePath + "/media")
+        self.mediaBox = MediaBox(basePath: basePath + "/media", isMainProcess: removeDatabaseOnError)
     }
     
     public func transaction<T>(ignoreDisabled: Bool = false, _ f: @escaping (AccountManagerModifier<Types>) -> T) -> Signal<T, NoError> {
